adjust clusterID based on version
