+++ packages/tests/src/utils/generate_test_data.ts
@@ -1,7 +1,5 @@
 import { createDecoder, createEncoder, Decoder, Encoder } from "@waku/core";
 
-import { DOCKER_IMAGE_NAME } from "../lib/service_node";
-
 // Utility to generate test data for multiple topics tests.
 export function generateTestData(topicCount: number): {
   contentTopics: string[];
@@ -21,20 +19,4 @@ export function generateTestData(topicCount: number): {
     encoders,
     decoders
   };
-}
-
-// Utility to add test conditions based on nwaku/go-waku versions
-export function isNwakuAtLeast(requiredVersion: string): boolean {
-  const versionRegex = /(?:v)?(\d+\.\d+(?:\.\d+)?)/;
-  const match = DOCKER_IMAGE_NAME.match(versionRegex);
-
-  if (match) {
-    const version = match[0].substring(1); // Remove the 'v' prefix
-    return (
-      version.localeCompare(requiredVersion, undefined, { numeric: true }) >= 0
-    );
-  } else {
-    // If there is no match we assume that it's a version close to master so we return True
-    return true;
-  }
 }
